Rename Modal component
